feat: Add scripts.ts, moved logic from it to helpers Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
