+++ typescript/vrotsc/src/compiler/transformers/helpers/README.md
@@ -1,0 +1,13 @@
+# What are these helpers?
+
+## node.ts
+
+Contains a set of functions to work with TypeScript nodes.
+The node helper does not modify the AST.
+It is not specific to Aria Orchestrator.
+
+## source.ts
+
+Contains a set of functions to work with TypeScript source files.
+The source helper can modify the AST.
+It is not specific to Aria Orchestrator.
